Changes - Special attack of npc - Refund system - Fixed fire breath - New commands - Boosted stats changes - Inferno cave entrance

+++ game-server/src/main/java/net/dodian/jobs/impl/NpcProcessor.java
@@ -2,12 +2,12 @@ package net.dodian.jobs.impl;
 
 import net.dodian.uber.game.Constants;
 import net.dodian.uber.game.Server;
+import net.dodian.uber.game.model.entity.Entity;
 import net.dodian.uber.game.model.entity.npc.Npc;
 import net.dodian.uber.game.model.entity.player.Client;
 import net.dodian.uber.game.model.entity.player.PlayerHandler;
 import net.dodian.uber.game.model.player.packets.outgoing.SendMessage;
 import net.dodian.uber.game.party.Balloons;
-import net.dodian.uber.game.model.UpdateFlag;
 import net.dodian.utilities.Misc;
 import net.dodian.utilities.Utils;
 import org.quartz.DisallowConcurrentExecution;
@@ -34,19 +34,21 @@ public class NpcProcessor implements Job {
              npc.setVisible(false);
              npc.drop();
              Client p = npc.getTarget(false);
+             npc.removeEnemy(p);
             /* Jad loot table, up to 5 players */
             if(npc.getId() == 3127)
-                for(int i = 0; i < 4 && p != null; i++) {
-                    npc.removeEnemy(p);
+                for(int i = 1; i <= 4 && !npc.getDamage().isEmpty(); i++) { //4 more players if damage not empty.
                     p = npc.getTarget(false);
                     if(p != null) {
-                        double chance = (npc.getDamage().get(p) / (double)npc.getMaxHealth()) * 100;
+                        double chance = (0.1 + (npc.getDamage().get(p) / (double)npc.getMaxHealth())) * 100; //10% + your damage profile!
                         double rate = Misc.chance(100000) / 1000D;
-                        if(rate <= chance) {
+                        if(chance - 10 >= 5 && rate <= chance) {
                             npc.drop();
                             p.send(new SendMessage("You managed to roll for the loot!"));
-                        } else p.send(new SendMessage("Unlucky! Better luck next time."));
+                        } else if(chance - 10 < 5) p.send(new SendMessage("You were not eligible for the drop!"));
+                        else p.send(new SendMessage("Unlucky! Better luck next time."));
                     }
+                    npc.removeEnemy(p); //Need to remove the enemy if we done with the check values!
                 }
         }
             if (!npc.alive && !npc.visible && (now - (npc.getDeathTime() + npc.getTimeOnFloor()) >= (npc.getRespawn() * 1000L))) {
